SEO-AUD-V1/requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
python-docx>=1.1.0
